Add the full JoimFW
Changed this repository to a single commit.
The past states are no longer relevant to anyone or anything.
This repository has been means to maintain and use code in a decentralised way, and we don't really need backwards compatibility.
Additionally, this commit reflects better the status of the working framework without third parties code.